+++ assess_environment.h
@@ -2,7 +2,7 @@
 #define ASSESS_ENVIRONMENT_H
 
 uint16_t get_line_position(void);
-int get_line_number(void);
+uint16_t get_line_number(void);
 void assess_environment_start(void);
 //int process_line_number(void);
 //uint16_t extract_line_width(uint8_t *buffer);
